Reorganize lookup / ctl modules
